requirements.txt
asgiref==3.3.1
autopep8==1.5.6
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.7
django-cors-headers==3.7.0
django-environ==0.4.5
django-heroku==0.3.1
django-import-export==2.5.0
djangorestframework==3.12.4
drf-yasg==1.20.0
et-xmlfile==1.0.1
gunicorn==20.1.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
joblib==1.0.1
MarkupPy==1.14
MarkupSafe==1.1.1
nltk==3.5
NRCLex==3.0.0
numpy==1.19.5
odfpy==1.4.1
openpyxl==3.0.7
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
psycopg2==2.9.1
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
tablib==3.0.0
textblob==0.15.3
threadpoolctl==2.1.0
toml==0.10.2
tqdm==4.60.0
uritemplate==3.0.1
urllib3==1.26.4
whitenoise==5.3.0
xlrd==2.0.1
xlwt==1.3.0
